requirements.txt
beautifulsoup4
python-docx
lxml
pillow~=11.3.0